% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_create_pTable.R
\name{create_num_ptable}
\alias{create_num_ptable}
\title{Create a ptable suitable for numerical variables}
\usage{
create_num_ptable(D, V, js = 0, pstay = NULL, optim = 1,
  mono = TRUE, step = 2, icat = NULL, type = "all",
  label = paste0("D", D, "V", V * 100), monitoring = FALSE,
  debugging = FALSE)
}
\arguments{
\item{D}{perturbation parameter for maximum noise/perturbation (scalar or vector)}

\item{V}{perturbation parameter for variance (scalar)}

\item{js}{treshold value for blocking of small frequencies (i.e. the perturbation
will not produce positive cell values that are equal to or smaller than
the treshold value).}

\item{pstay}{optional parameter to set the probability (0 < p < 1) of
an original frequency to remain unperturbed: NA (default) no preset
probability (i.e. produces the maximum entropy solution)}

\item{optim}{optimization parameter: \code{1} standard approach (default)}

\item{mono}{(logical) vector specifying optimization parameter for
monotony condition}

\item{step}{(integer) step width}

\item{icat}{(integer) categorized original frequencies i}

\item{type}{(character) indicator for the extra column 'type' used for
magnitude tables: 'even', 'odd' or 'all' (default)}

\item{label}{(character) label of the Output}

\item{monitoring}{(logical) output monitoring on/off}

\item{debugging}{(logical) debug monitoring on/off}
}
\value{
an object of \linkS4class{ptable}
}
\description{
\code{\link[=create_num_ptable]{create_num_ptable()}} is a wrapper for \code{\link[=pt_create_pTable]{pt_create_pTable()}} allowing
only to specify the relevant options for numerical variables.
}
\examples{
# ptable for frequency tables 
# old way
params_cnts <- pt_create_pParams(D = 5, V = 3, js = 2, label = "test1")
pt_create_pTable(params = params_cnts)

# new, direct way
create_cnt_ptable(D = 5, V = 3, js = 2, label = "test2")

# ptable for magnitude tables
# old way
params_nums <- pt_create_pParams(D=5, V=2, table="nums", step=4, icat=c(1, 3, 5))
pt_create_pTable(params = params_nums)

# new way
create_num_ptable(D = 5, V = 2, step = 4, icat = c(1, 3, 5))
}
\seealso{
\code{\link[=pt_create_pTable]{pt_create_pTable()}}, \code{\link[=pt_create_pParams]{pt_create_pParams()}}
}
\author{
Bernhard Meindl, \email{bernhard.meindl@statistik.gv.at}
}
