% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pt_check}
\alias{pt_check}
\title{Check the constraint of the ptable}
\usage{
pt_check(ptab)
}
\arguments{
\item{ptab}{a \code{data.table} or an an object of \linkS4class{ptable} generated with \code{\link[=create_cnt_ptable]{create_cnt_ptable()}}.}
}
\value{
a data.table object
}
\description{
\code{\link[=pt_check]{pt_check()}} checks the constraints of the ptable
}
\examples{
# create ptable 
ptab1 <- create_cnt_ptable(D = 5, V = 3, js = 2, label = "test2")

# check ptable
pt_check(ptab1)

}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{check}
\keyword{ptable}
