% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_examples.R
\name{pt_ex_cnts}
\alias{pt_ex_cnts}
\title{A quick ptable that can be used in various examples}
\usage{
pt_ex_cnts()
}
\value{
an object of \linkS4class{ptable}
}
\description{
\code{\link[=pt_ex_cnts]{pt_ex_cnts()}} returns a perturbation table object from
\code{\link[=pt_create_pTable]{pt_create_pTable()}} with some default parameters. This is useful
for quickly creating ptables to demonstate usage in other tools.
}
\examples{
ptab <- pt_ex_cnts()
plot(ptab, type="t")
}
