% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_vignette.R
\name{pt_vignette}
\alias{pt_vignette}
\title{Vignette}
\usage{
pt_vignette()
}
\value{
a browser windows/tab with showing the vignette
}
\description{
Starts the package vignette that gets you started with the 
package
}
\examples{
\dontrun{
#pt_vignette()
}
}
