% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_classes.R
\docType{class}
\name{ptable-class}
\alias{ptable-class}
\title{An S4 class to represent perturbation table}
\description{
An S4 class to represent perturbation table
}
\section{Slots}{

\describe{
\item{\code{tMatrix}}{(matrix) transition matrix with perturbation probabilities}

\item{\code{pClasses}}{(numeric) numeric classes}

\item{\code{pTable}}{(data.table) perturbation table with probabilities}

\item{\code{empResults}}{(data.table) ...}

\item{\code{pParams}}{a \code{\linkS4class{ptable_params}} object}

\item{\code{tStamp}}{(character) ...}

\item{\code{type}}{(character)}

\item{\code{table}}{(character) type of table: frequency counts (cnts) or magnitude (nums)}
}}

