% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptable.R
\name{ptable}
\alias{ptable}
\title{ptable}
\usage{
ptable()
}
\description{
Perturbation Table Visualizer
}
\examples{
\dontrun{
ptable()
}

}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{dashboard,}
\keyword{flexdashboard,}
\keyword{perturbation}
\keyword{shiny}
\keyword{table,}
\keyword{visualisation,}
