% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_methods.R
\name{plot}
\alias{plot}
\title{Method to plot results}
\usage{
plot(obj, type = "d", file = NULL, ...)
}
\arguments{
\item{obj}{an object of class \code{\linkS4class{ptable}}}

\item{type}{(character) type of graph: distribution "d" (standard), perturbation panel ("p"), transition matrix "t"}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to as pdf}

\item{...}{additional parameters passed to methods}
}
\description{
Function to plot the perturbation panel using the R-package \code{\link{ggplot2}}.
}
\examples{

ptab <- create_cnt_ptable(D = 5, V = 2, label = "Example")

# Distribution Plot of the Noise
plot(ptab, type = "d")

# Perturbations Panel of the Noise
plot(ptab, type = "p")

# Transition Matrix
plot(ptab, type = "t")

\dontrun{
## Export result
plot(ptab, type ="t", file = "example_tMatrix.pdf")
}

}
