% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_methods.R
\name{plot}
\alias{plot}
\title{Method to plot results}
\usage{
plot(obj, type = "d", file = NULL, ...)
}
\arguments{
\item{obj}{an object of class \code{\linkS4class{ptable}}}

\item{type}{(character) type of graph: distribution "d" (standard), perturbation panel ("p"), transition matrix "t"}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to as pdf}

\item{...}{additional parameters passed to methods}
}
\description{
Method to plot results
}
\examples{
# Simple Example
ptab <- create_cnt_ptable(D = 2, V = 1.08, js = 1, mono = FALSE, label = "Example")
plot(ptab, type = "d")

\dontrun{
## Export result
plot(ptab, type = "d", file = "graph.pdf")
}

}
