% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_plot.R
\name{plot}
\alias{plot}
\title{Plot the results of the perturbation table generator}
\usage{
plot(obj, type = "d", file = NULL, ...)
}
\arguments{
\item{obj}{an object of class \code{\linkS4class{ptable}}}

\item{type}{(character) type of graph: distribution "d" (standard),
perturbation panel ("p"), transition matrix "t"}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to
as pdf}

\item{...}{additional parameters passed to methods}
}
\description{
\code{\link[=plot]{plot()}} makes it easy to visualize the results of the created
ptable-object that has been created by \code{\link[=create_cnt_ptable]{create_cnt_ptable()}},
\code{\link[=create_cnt_ptable]{create_cnt_ptable()}} or \code{\link[=modify_cnt_ptable]{modify_cnt_ptable()}}.
}
\examples{

ptab <- create_cnt_ptable(D = 5, V = 2, label = "Example")
ptab_mod <- modify_cnt_ptable(ptab, threshold = 0.3, seed = 5432)

# Distribution Plot of the Noise
plot(ptab, type = "d")

# Perturbations Panel of the Noise
plot(ptab, type = "p")
plot(ptab_mod, type = "p")

# Transition Matrix
plot(ptab, type = "t")

\dontrun{
## Export result
plot(ptab, type ="t", file = "example_tMatrix.pdf")
}

}
\author{
Tobias Enderle
}
\keyword{plot}
