% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_create_pParams.R
\name{pt_create_pParams}
\alias{pt_create_pParams}
\title{pt_create_pParams}
\usage{
pt_create_pParams(D, V, js = 0, pstay = NULL, optim = 1, mono = TRUE,
  epsilon = 0.0000001, label = paste("D", D, "V", V * 100, sep = ""),
  pTableSize = 70)
}
\arguments{
\item{D}{perturbation parameter for maximum noise/perturbation (scalar or vector)}

\item{V}{perturbation parameter for variance (scalar)}

\item{js}{treshold value for blocking of small frequencies (i.e. the perturbation will not produce positive cell values that are equal to or smaller than the treshold value).}

\item{pstay}{optional parameter to set the probability (0 < p < 1) of an original frequency to remain unperturbed: NA (default) no preset probability (i.e. produces the maximum entropy solution)}

\item{optim}{optimization parameter: \code{1} standard approach (default)}

\item{mono}{(logical) vector specifying optimization parameter for monotony condition}

\item{epsilon}{(double)}

\item{label}{(character) label of the Output}

\item{pTableSize}{(number) defining the number of required columns if abs-formatted ptable is chosen as argument in \code{\link{pt_create_pTable}}}
}
\value{
an object of \code{\linkS4class{ptable_params}}
}
\description{
creates the required input for \code{\linkS4class{ptable}}.
}
\examples{
pt_create_pParams(D=5, V=2, js=2, pstay=0.5, optim=1, mono=TRUE)

}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{data}
