% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_create_pParams.R
\name{pt_create_pParams}
\alias{pt_create_pParams}
\title{pt_create_pParams}
\usage{
pt_create_pParams(D, V, js = 0, pstay = NULL, optim = 1, mono = TRUE,
  epsilon = 1e-07, label = paste("D", D, "V", V * 100, sep = ""))
}
\arguments{
\item{D}{perturbation parameter for maximum perturbation (scalar or vector)}

\item{V}{perturbation parameter for variance (scalar)}

\item{js}{treshold value for blocking of small frequencies (i.e. there won't occur positive target frequencies below the treshold value)
Target frequencies are defined by ...}

\item{pstay}{optional parameter to set}

\item{optim}{optimization parameter: \code{1} standard approach (default)}

\item{mono}{(logical) vector specifying optimization parameter for monotony condition}

\item{epsilon}{(double)}

\item{label}{(character) label of the Output}
}
\value{
an object of \code{\linkS4class{ptable_params}}
}
\description{
creates the required input for \code{\linkS4class{ptable}}.
}
\examples{
pt_create_pParams(D=5, V=2, js=2, pstay=0.5, optim=1, mono=TRUE)

}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{data}
