% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_create_pTable.R
\name{pt_create_pTable}
\alias{pt_create_pTable}
\title{pt_create_pTable}
\usage{
pt_create_pTable(params, monitoring = FALSE, debugging = FALSE)
}
\arguments{
\item{params}{an object of class \code{\linkS4class{ptable_params}}
generated with \code{\link{pt_create_pParams}}}

\item{monitoring}{(logical) output monitoring on/off}

\item{debugging}{(logical) debug monitoring on/off}
}
\value{
an object of \code{\linkS4class{ptable}}
}
\description{
produces perturbation table that is needed to add noise to statistical frequency or magnitude tables. The perturbation probabilities are constructed given the following constraints:
\itemize{
\item Unbiasedness of the noise
\item Fixed noise variance
\item Transition probabilities are between zero and one and the sum up to 1
\item Perturbations will not produce negaive cell values or positive cell values equal to or less than a specific threshold value
\item The absolute value of any perturbation is less than a specific integer value (i.e. the maxiumum noise)
}
}
\details{
For more information, see the vignette using \code{pt_vignette()}.
}
\examples{
# ptable for frequency tables
params_cnts <- pt_create_pParams(D=5, V=3, js=2, label="test")
pt_create_pTable(params=params_cnts)

# ptable for magnitude tables
params_nums <- pt_create_pParams(D=5, V=2, table="nums", step=4, icat=c(1,3,5))
pt_create_pTable(params=params_nums)

}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{perturbation}
\keyword{sdc}
\keyword{table,}
