% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_create_pTable.R
\name{pt_create_pTable}
\alias{pt_create_pTable}
\title{pt_create_pTable}
\usage{
pt_create_pTable(params, type, monitoring = FALSE)
}
\arguments{
\item{params}{an object of class \code{\linkS4class{ptable_params}}
generated with \code{\link{pt_create_pParams}}}

\item{type}{(character) type of pTable (either 'abs' or 'destatis')}

\item{monitoring}{(logical) debug monitoring on/off}
}
\value{
an object of \code{\linkS4class{ptable}}
}
\description{
produces perturbation table that is needed to add noise to statistical frequency tables. The perturbation probabilities are constructed given the following constraints:
\enumerate{
\item Zero mean
\item Constant variance
\item Probabilities are between zero and one
\item 
\item probabilities sum up to 1
}
}
\examples{
(params <- pt_create_pParams(D=5, V=2, js=2, label="test"))
\dontrun{
pt_create_pTable(params=params, type="destatis")
}
}
\author{
Tobias Enderle, \email{tobias.enderle@destatis.de}
}
\keyword{perturbation}
\keyword{sdc}
\keyword{table,}
