% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_export.R
\name{pt_export}
\alias{pt_export}
\title{Export ptables as a csv-file}
\usage{
pt_export(..., file, SDCtool = "TauArgus")
}
\arguments{
\item{...}{1 or 2 input object of class \code{\linkS4class{ptable}}}

\item{file}{(character) filename}

\item{SDCtool}{(character) either "TauArgus" or "SAS"}
}
\description{
Function to export perturbation table to Tau-Argus, SAS or
any other CKM tool (as csv-file).
}
\examples{
ptab <- create_cnt_ptable(D = 5, V = 3, js = 2, label = "test")
\dontrun{
pt_export(ptab, file = "Test", SDCtool = "TauArgus")
}
}
\author{
Tobias Enderle
}
\keyword{export}
