% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_export.R
\name{pt_export}
\alias{pt_export}
\title{pt_export}
\usage{
pt_export(input, file)
}
\arguments{
\item{input}{an object of class \code{\linkS4class{ptable}}}

\item{file}{(character) filename}
}
\description{
Function to export ptable (csv).
}
\examples{
params <- pt_create_pParams(D=5, V=3, js=2, label="test")
ptable_destatis <- pt_create_pTable(params=params, type="destatis")
\dontrun{
pt_export(ptable_destatis,file="Test")
}

}
\author{
Tobias Enderle
}
\keyword{export}
