% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_export.R
\name{pt_export}
\alias{pt_export}
\title{pt_export}
\usage{
pt_export(..., file, SDCtool = "TauArgus")
}
\arguments{
\item{...}{1 or 2 input object of class \code{\linkS4class{ptable}}}

\item{file}{(character) filename}

\item{SDCtool}{(character) either "TauArgus" or "SAS"}
}
\description{
Function to export perturbation table to Tau-Argus or SAS (as csv-file).
}
\examples{
params <- pt_create_pParams(D=5, V=3, js=2, label="test")
ptab <- pt_create_pTable(params=params)
\dontrun{
pt_export(ptab,file="Test", SDCtool="TauArgus")
}

}
\author{
Tobias Enderle
}
\keyword{export}
