% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_plot.R
\name{fifi_plot}
\alias{fifi_plot}
\title{pt_plot}
\usage{
fifi_plot(pert_table, ylimit = c(-0.05, 0.95))
}
\arguments{
\item{pert_table}{an object of class \code{\linkS4class{ptable}}}

\item{ylimit}{(numeric) vector with limits of y-axis (for probabilities)}
}
\description{
Function to plot the perturbation table using the R-package \code{\link{lattice}}.
}
\examples{
# Simple Example
params <- pt_create_pParams(D=5, V=2, js=2, label="Example")
ptable_destatis <- pt_create_pTable(params=params, type="destatis")
fifi_plot(ptable_destatis)

# Export result
pdf(paste("graph.pdf", sep=""))
  print(fifi_plot(ptable_destatis))
dev.off()

}
\author{
Tobias Enderle
}
\keyword{plot}
