% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_plot_pD.R
\name{pt_plot_pD}
\alias{pt_plot_pD}
\title{pt_plot_pD}
\usage{
pt_plot_pD(pert_table, ylimit = c(-0.05, 0.95), file = NULL)
}
\arguments{
\item{pert_table}{an object of class \code{\linkS4class{ptable}}}

\item{ylimit}{(numeric) vector with limits of y-axis (for probabilities)}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to as pdf}
}
\description{
Function to plot the ditribution of the perturbation values using the R-package \code{\link{lattice}}.
}
\examples{
# Simple Example
params <- pt_create_pParams(D=5, V=2, label="Example")
ptable_destatis <- pt_create_pTable(params=params)
plot(ptable_destatis, type="d")

\dontrun{
## Export result
plot(ptable_destatis, type="d", file="graph.pdf")
}
}
\author{
Tobias Enderle
}
\keyword{plot}
