% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_plot_pPanel.R
\name{pt_plot_pPanel}
\alias{pt_plot_pPanel}
\title{pt_plot_pPanel}
\usage{
pt_plot_pPanel(pert_table, file = NULL)
}
\arguments{
\item{pert_table}{an object of class \code{\linkS4class{ptable}}}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to as pdf}
}
\description{
Function to plot the perturbation panel using the R-package \code{\link{ggplot}}.
}
\examples{
# Simple Example
ptab <- create_cnt_ptable(D = 5, V = 2, label = "Example")
plot(ptab, type="p")

\dontrun{
## Export result
plot(ptab, type = "p", file = "example_pPanel.png")
}
}
\author{
Tobias Enderle
}
\keyword{plot}
