% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_plot_tMatrix.R
\name{pt_plot_tMatrix}
\alias{pt_plot_tMatrix}
\title{pt_plot_tMatrix}
\usage{
pt_plot_tMatrix(pert_table, file = NULL)
}
\arguments{
\item{pert_table}{an object of class \code{\linkS4class{ptable}}}

\item{file}{if not \code{NULL}, a path to a file where the graph is saved to as pdf}
}
\description{
Function to plot the transition matrix using the R-package \code{\link{ggplot}}.
}
\examples{
# Simple Example
params <- pt_create_pParams(D=5, V=2, label="Example")
ptable_destatis <- pt_create_pTable(params=params, type="destatis")
fifi_plot(ptable_destatis, type="t")

\dontrun{
## Export result
fifi_plot(ptable_destatis, type="t", file="example_tMatrix.pdf")
}
}
\author{
Tobias Enderle
}
\keyword{plot}
