% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_classes.R
\docType{class}
\name{ptable-class}
\alias{ptable-class}
\title{An S4 class to represent perturbation table}
\description{
An S4 class to represent perturbation table
}
\section{Slots}{

\describe{
\item{\code{pMatrix}}{(matrix) perturbation matrix with probabilites}

\item{\code{pClasses}}{(numeric) numeric classes}

\item{\code{pTable}}{(data.table) perturbation table with probabilities}

\item{\code{dFrame}}{(data.frame) for visualizing issues (to be done in upcoming releases)}

\item{\code{empResults}}{(data.table) ...}

\item{\code{pParams}}{a \code{\linkS4class{ptable_params}} object}

\item{\code{tStamp}}{(character) ...}

\item{\code{type}}{(character) specifying the type of pTable (either 'abs' or 'destatis')}
}}

