% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_classes.R
\docType{class}
\name{ptable_params-class}
\alias{ptable_params-class}
\title{An S4 class to represent perturbation parameters}
\description{
An S4 class to represent perturbation parameters
}
\section{Slots}{

\describe{
\item{\code{D}}{(integer) parameter for maximum perturbation / noise}

\item{\code{V}}{(numeric) parameter for perturbation variance}

\item{\code{js}}{(integer) parameter for original counts not to be perturbed}

\item{\code{ncat}}{(integer) number of perturbation classes}

\item{\code{pstay}}{numeric vector specifying parameter for non-perturbation}

\item{\code{optim}}{(integer) specifying optimization parameter for optimization 
function}

\item{\code{mono}}{(logical) vector specifying optimization parameter for monotony 
condition}

\item{\code{label}}{(character) label for output}

\item{\code{icat}}{(integer) categorized original frequencies i}

\item{\code{table}}{(character) type of table: frequency counts (\code{cnts}) or 
magnitude (\code{nums})}

\item{\code{step}}{(integer) step}

\item{\code{type}}{(character) indicator for the extra column 'type' used for
magnitude tables: \code{'even'}, \code{'odd'} or \code{'all'}}
}}

